requirements.txt
Django>=5.0.1
Pillow>=10.0.0
markdown>=3.5
djangorestframework>=3.16.1
drf-spectacular>=0.29.0
google-genai>=1.30.0
playwright>=1.50.0
